Moved LoginViewModel.kt to viewmodel package.
