+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentormanager/ui/viewmodels/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.peculiaruc.alc_mmsystem_mentormanager.ui.fragments
+package com.peculiaruc.alc_mmsystem_mentormanager.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
